dockerfile_api/requirements.txt
cassandra-driver==3.24.0
uvicorn==0.13.3
fastapi==0.63.0